% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dg_closest_res_to_area}
\alias{dg_closest_res_to_area}
\title{Determine resolution based on desired area}
\usage{
dg_closest_res_to_area(
  dggs,
  area,
  round = "nearest",
  show_info = TRUE,
  metric = TRUE
)
}
\arguments{
\item{dggs}{A dggs object from dgconstruct()}

\item{area}{The desired area of the grid's cells}

\item{round}{What direction to search in. Must be nearest, up, or down.}

\item{show_info}{Print the area, spacing, and CLS of the chosen resolution.}

\item{metric}{Whether input and output should be in metric (TRUE) or
imperial (FALSE)}
}
\value{
A number representing the grid resolution
}
\description{
Determine an appropriate grid resolution based on a desired
                 cell area.
}
\examples{
library(dggridR)
dggs <- dgconstruct(res=20)
res  <- dg_closest_res_to_area(dggs,1)
dgsetres(dggs,res)

}
