% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dg_process_polydata}
\alias{dg_process_polydata}
\title{Load a KML file}
\usage{
dg_process_polydata(polydata)
}
\arguments{
\item{polydata}{Polygons generated by dggrid. These will be converted.}
}
\value{
Returns an sf object.
}
\description{
Convert data from internal dggrid functions into something
                 useful: an sp object or a data frame
}
\keyword{internal}
