% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dg_transform_for_output}
\alias{dg_transform_for_output}
\title{Transform a dggs object's fields to a form suitable for output}
\usage{
dg_transform_for_output(dggs)
}
\arguments{
\item{dggs}{A dggs object from dgconstruct().}
}
\description{
A dggs object may contain variables that are named for user
       friendliness. This function removes the friendliness. It's enemies
       from here on out.
}
