% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwrapper.R
\name{dgGEO_to_PROJTRI}
\alias{dgGEO_to_PROJTRI}
\title{Convert from GEO to PROJTRI}
\usage{
dgGEO_to_PROJTRI(dggs, in_lon_deg, in_lat_deg)
}
\arguments{
\item{lon_deg}{Vector of longitude, in degrees}

\item{lat_deg}{Vector of latitude, in degrees}
}
\value{
Returns a dggs object which can be passed to other dggridR
            functions
}
\description{
Uses a discrete global grid system to convert between GEO
            and PROJTRI (see vignette for details)
}
\examples{
library(dggridR)
dggs <- dgconstruct(res=20)

GEO_to_PROJTRI(dggs, in_lon_deg, in_lat_deg)

}
