% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dginfo}
\alias{dginfo}
\title{Print info about a dggs object to the screen}
\usage{
dginfo(dggs)
}
\arguments{
\item{dggs}{A dggs object from dgconstruct()}
}
\value{
No return. All info is printed to the screen.
}
\description{
dggs objects have many settings. This returns all of them, along
         with info about the grid being specified.
}
\examples{
library(dggridR)
dginfo(dgconstruct(res=20))

}
