% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwrapper.R
\name{dgPROJTRI_to_Q2DI}
\alias{dgPROJTRI_to_Q2DI}
\title{Convert from PROJTRI to Q2DI}
\usage{
dgPROJTRI_to_Q2DI(dggs, in_tnum, in_tx, in_ty)
}
\arguments{
\item{tnum}{Vector of triangle numbers}

\item{tx}{Vector of triangle x values}

\item{ty}{Vector of triangle y values}
}
\value{
Returns a dggs object which can be passed to other dggridR
            functions
}
\description{
Uses a discrete global grid system to convert between PROJTRI
            and Q2DI (see vignette for details)
}
\examples{
library(dggridR)
dggs <- dgconstruct(res=20)

PROJTRI_to_Q2DI(dggs, in_tnum, in_tx, in_ty)

}
