% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwrapper.R
\name{dgPROJTRI_to_SEQNUM}
\alias{dgPROJTRI_to_SEQNUM}
\title{Convert from PROJTRI to SEQNUM}
\usage{
dgPROJTRI_to_SEQNUM(dggs, in_tnum, in_tx, in_ty)
}
\arguments{
\item{in_tnum}{Vector of triangle numbers}

\item{in_tx}{Vector of triangle x values}

\item{in_ty}{Vector of triangle y values}
}
\value{
Returns a dggs object which can be passed to other dggridR
            functions
}
\description{
Uses a discrete global grid system to convert between PROJTRI
            and SEQNUM (see vignette for details)
}
\examples{
library(dggridR)
dggs <- dgconstruct(res=20)

dgPROJTRI_to_SEQNUM(dggs, in_tnum, in_tx, in_ty)

}
