% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwrapper.R
\name{dgQ2DD_to_SEQNUM}
\alias{dgQ2DD_to_SEQNUM}
\title{Convert from Q2DD to SEQNUM}
\usage{
dgQ2DD_to_SEQNUM(dggs, in_quad, in_qx, in_qy)
}
\arguments{
\item{quad}{Vector of quad numbers}

\item{qx}{Vector of quadrant x values}

\item{qy}{Vector of quadrant y values}
}
\value{
Returns a dggs object which can be passed to other dggridR
            functions
}
\description{
Uses a discrete global grid system to convert between Q2DD
            and SEQNUM (see vignette for details)
}
\examples{
library(dggridR)
dggs <- dgconstruct(res=20)

Q2DD_to_SEQNUM(dggs, in_quad, in_qx, in_qy)

}
