% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwrapper.R
\name{dgQ2DI_to_Q2DD}
\alias{dgQ2DI_to_Q2DD}
\title{Convert from Q2DI to Q2DD}
\usage{
dgQ2DI_to_Q2DD(dggs, in_quad, in_i, in_j)
}
\arguments{
\item{quad}{Vector of quad numbers}

\item{i}{Vector of quadrant i values}

\item{j}{Vector of quadrant j values}
}
\value{
Returns a dggs object which can be passed to other dggridR
            functions
}
\description{
Uses a discrete global grid system to convert between Q2DI
            and Q2DD (see vignette for details)
}
\examples{
library(dggridR)
dggs <- dgconstruct(res=20)

Q2DI_to_Q2DD(dggs, in_quad, in_i, in_j)

}
