% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dgrectgrid}
\alias{dgrectgrid}
\title{Return the coordinates constituting the boundary of cells within a
         specified region}
\usage{
dgrectgrid(
  dggs,
  minlat = -1,
  minlon = -1,
  maxlat = -1,
  maxlon = -1,
  cellsize = 0.1,
  ...
)
}
\arguments{
\item{dggs}{A dggs object from dgconstruct()}

\item{minlat}{Minimum latitude of region of interest}

\item{minlon}{Minimum longitude of region of interest}

\item{maxlat}{Maximum latitude of region of interest}

\item{maxlon}{Maximum longitude of region of interest}

\item{cellsize}{Distance, in degrees, between the sample points used to
generate the grid. Small values yield long generation times
while large values may omit cells.}

\item{\dots}{Further arguments passed to \code{\link{dgcellstogrid}}.}
}
\value{
Returns an sf object.
        If \code{!is.na(savegrid)}, returns a filename.
}
\description{
Note: This may generate odd results for very large
                 rectangles, because putting rectangles on spheres is
                 weird... as you should know, if you're using this package.
}
\examples{
library(dggridR)
dggs <- dgconstruct(spacing=1000,metric=FALSE,resround='down')

#Get grid cells for the conterminous United States
grid <- dgrectgrid(dggs,
               minlat=24.7433195, minlon=-124.7844079,
               maxlat=49.3457868, maxlon=-66.9513812)
head(grid)

}
