% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dgrun}
\alias{dgrun}
\title{A generic function for running dggrid and returning values from it}
\usage{
dgrun(dggs, clean = TRUE, check = TRUE, has_output_file = TRUE)
}
\arguments{
\item{dggs}{A dggs object from dgconstruct()}

\item{clean}{If TRUE, delete the temporary files used/produced by dggrid. TODO}

\item{check}{If TRUE, raise a stop signal if dggrid doesn't run successfully.}

\item{has_output_file}{If TRUE, look for output file. If FALSE, capture and
return stdout and stderr.}
}
\value{
If \code{has_output_file} is TRUE, a data frame is returned.
         The calling function is responsible for understanding this frame.
         If \code{has_output_file} is FALSE, the stdout and stderr of dggrid
         are captured and returned as a character vector.
}
\description{
A discrete global grid maps lat-long points to particular cells.
         These cells are uniquely numbered, for a given resolution, from
         1 to some maximum number. Cell numbers may be reused from one
         resolution to the next.
}

