% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dgsavegrid}
\alias{dgsavegrid}
\title{Saves a generated grid to a shapefile}
\usage{
dgsavegrid(grid, shpfname)
}
\arguments{
\item{grid}{Grid to be saved}

\item{shpfname}{File to save the grid to}
}
\value{
The filename the grid was saved to
}
\description{
Saves a generated grid to a shapefile
}
\keyword{internal}
