% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwrapper.R
\name{dgSEQNUM_to_PLANE}
\alias{dgSEQNUM_to_PLANE}
\title{Convert from SEQNUM to PLANE}
\usage{
dgSEQNUM_to_PLANE(dggs, in_seqnum)
}
\arguments{
\item{in_seqnum}{Globally unique number identifying the surface polygon}
}
\value{
Returns a dggs object which can be passed to other dggridR
            functions
}
\description{
Uses a discrete global grid system to convert between SEQNUM
            and PLANE (see vignette for details)
}
\examples{
library(dggridR)
dggs <- dgconstruct(res=20)

dgSEQNUM_to_PLANE(dggs, in_seqnum)

}
