% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\docType{data}
\name{dg_env}
\alias{dg_env}
\title{Control global aspects of the dggridR package}
\format{An object of class \code{environment} of length 1.}
\usage{
dg_env
}
\description{
This environment is used to control global features of the dggridR
       package. At the moment the only option is 'dg_debug' which, when set
       to TRUE provides extensive outputs useful for tracking down bugs.
}
\keyword{datasets}
