% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dg_exe_path}
\alias{dg_exe_path}
\title{Get path to dggrid executable}
\usage{
dg_exe_path()
}
\value{
A string representing the path to the dggrid executable.
}
\description{
Returns a path to the dggrid executable. Used for running stuff.
}
