% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dg_process_kml}
\alias{dg_process_kml}
\title{Load a KML file}
\usage{
dg_process_kml(kmlfile, frame)
}
\arguments{
\item{kmlfile}{A KML file generated by dgrectgrid() or dgearthgrid()}

\item{frame}{If TRUE, return a data frame suitable for ggplot plotting.
If FALSE, return an OGR poly object}
}
\value{
Returns a data frame or OGR poly object, as specified by \code{frame}
}
\details{
Clean up and/or convert a KML grid cell boundary file
}

