% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dg_process_polydata}
\alias{dg_process_polydata}
\title{Load a KML file}
\usage{
dg_process_polydata(polydata, frame, wrapcells)
}
\arguments{
\item{polydata}{Polygons generated by dggrid. These will be converted.}

\item{frame}{If TRUE, return a data frame suitable for ggplot plotting.
If FALSE, return an SpatialPolygons}

\item{wrapcells}{Cells which cross -180/180 degrees can present 
difficulties for plotting. Setting this TRUE will result in
cells with components in both hemispheres to be mapped
entirely to positive degrees (the Eastern hemisphere). As a
result, such cells will have components in the range
[180,360). Only used when \code{frame=TRUE}.}
}
\value{
Returns a data frame or OGR poly object, as specified by \code{frame}
}
\description{
Convert data from internal dggrid functions into something
                 useful: an sp object or a data frame
}
