% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dg_shpfname_south_africa}
\alias{dg_shpfname_south_africa}
\title{National border of South Africa}
\usage{
dg_shpfname_south_africa()
}
\value{
A filename of a shapefile containing the national border of South Africa
}
\description{
This variable points to a shapefile containing the national border
       of South Africa
}
