% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dgconstruct}
\alias{dgconstruct}
\title{Construct a discrete global grid system (dggs) object}
\usage{
dgconstruct(type = "ISEA3H", res = 9, precision = 7)
}
\arguments{
\item{type}{Type of grid to use. Options are: ISEA3H, ISEA4H, ISEA43H, 
ISEA4T, ISEA4D, FULLER3H, FULLER4H, FULLER43H, FULLER4T,
and FULLER4D. Default: ISEA3H}

\item{res}{Resolution. Must be in the range [0,35]. Larger values represent
finer resolutions. Appropriate resolutions can be found with
dg_closest_res_to_area(), dg_closest_res_to_spacing(), and
dg_closest_res_to_cls(). Default is 9, which corresponds to a
cell area of ~2600 sq km and a cell spacing of ~50 km.
Default: 9.}

\item{precision}{Round output to this number of decimal places. Must be in
the range [0,30]. Default: 7.}
}
\value{
Returns a dggs object which can be passed to other dggridR
            functions
}
\details{

}
\examples{
library(dggridR)
dggs <- dgconstruct(res=20)

}

