% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dgearthgrid}
\alias{dgearthgrid}
\title{Return the coordinates constituting the boundary of cells for the
         entire Earth}
\usage{
dgearthgrid(dggs, frame = TRUE)
}
\arguments{
\item{dggs}{A dggs object from dgconstruct()}

\item{frame}{If TRUE, return a data frame suitable for ggplot plotting.
If FALSE, return an OGR poly object}
}
\value{
Returns a data frame or OGR poly object, as specified by \code{frame}
}
\description{
Note: If you have a high-resolution grid this may take a
                 loooooong time to execute.
}
\examples{
library(dggridR)
dggs <- dgconstruct(res=20)
res  <- dg_closest_res_to_spacing(dggs,spacing=1000,round='down',metric=FALSE)
dggs <- dgsetres(dggs,res)
grid <- dgearthgrid(dggs,frame=TRUE)

}

