% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dgneighbors}
\alias{dgneighbors}
\title{Return neighboring cell IDs for each input cell}
\usage{
dgneighbors(dggs, cells)
}
\arguments{
\item{dggs}{A dggs object from \code{\link{dgconstruct}()}}

\item{cells}{Integer vector of cell sequence numbers (SEQNUM)}
}
\value{
A data frame with columns \code{seqnum} (the input cell) and
        \code{neighbor} (each adjacent cell ID).
}
\description{
For each cell ID in \code{cells}, returns the IDs of all
                 adjacent cells. Triangle grids are not supported.
}
\examples{
library(dggridR)
dggs <- dgconstruct(res=3)
nbrs <- dgneighbors(dggs, c(1, 2, 3))

}
