% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dgparent}
\alias{dgparent}
\title{Return parent cell ID for each input cell}
\usage{
dgparent(dggs, cells)
}
\arguments{
\item{dggs}{A dggs object from \code{\link{dgconstruct}()}.
The parent cells will be at \code{dggs\$res - 1}.}

\item{cells}{Integer vector of cell sequence numbers (SEQNUM)}
}
\value{
A data frame with columns \code{seqnum} (the input cell) and
        \code{parent} (the parent cell ID at resolution - 1).
}
\description{
For each cell ID in \code{cells} at the current grid
                 resolution, returns the ID of its parent cell at resolution
                 \code{dggs\$res - 1}. Only hexagonal grids are supported.
}
\examples{
library(dggridR)
dggs <- dgconstruct(res=4)
prnt <- dgparent(dggs, c(1, 2))

}
