% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dgrectgrid}
\alias{dgrectgrid}
\title{Return the coordinates constituting the boundary of cells within a
         specified region}
\usage{
dgrectgrid(dggs, minlat = -1, minlon = -1, maxlat = -1, maxlon = -1,
  frame = TRUE)
}
\arguments{
\item{dggs}{A dggs object from dgconstruct()}

\item{minlat}{Minimum latitude of region of interest}

\item{minlon}{Minimum longitude of region of interest}

\item{maxlat}{Maximum latitude of region of interest}

\item{maxlon}{Maximum longitude of region of interest}

\item{frame}{If TRUE, return a data frame suitable for ggplot plotting.
If FALSE, return an OGR poly object}
}
\value{
Returns a data frame or OGR poly object, as specified by \code{frame}
}
\details{
Note: This may generate odd results for very large
                 rectangles, because putting rectangles on spheres is
                 weird... as you should know, if you're using this package.
}
\examples{
library(dggridR)
dggs <- dgconstruct(spacing=1000,metric=FALSE,round='down')

#Get grid cells for the conterminous United States
grid <- dgrectgrid(dggs,minlat=24.7433195, minlon=-124.7844079, maxlat=49.3457868, maxlon=-66.9513812, frame=TRUE)

}

