% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggridR.R
\name{dgverify}
\alias{dgverify}
\title{Verify that a dggs object has appropriate values}
\usage{
dgverify(dggs)
}
\arguments{
\item{dggs}{The dggs object to be verified}
}
\value{
The function has no return value. A stop signal is raised if the
            object is misspecified
}
\details{
Verify that a dggs object has appropriate values
}
\examples{
library(dggridR)
dggs <- dgconstruct(res=20)
dgverify(dggs)
}

